refactor(control-api): extract dto from policies/mod.rs
